Run the bot forever
This adds a new decorator rightly named run_forever, and its test
function.

It should run and tell us exatly what is going on when it fails.
For now it only knows about IOError, and this should cover any
connection problem. it just prints the Exception message when
it's unknown.

It also allows the user to stop the bot with Ctrl+C.

Later we hope to have zero traceback and handle every problem
gracefully.

For the test function test_running_loop, we only have naive tests. We
should also measure the wait time to see if it's working correctly (see
TODO).