Rework matching algorithm
We now support banning tweets by screen_name and clients. This is useful
because we tend to find lots of tweets sent by bots. The main matching
function has been reworked to support that.

Also, the main matching approach changed to use a simplematcher first,
which proved a bit better(about 8% faster).