New watchdog/auto-reset feature
Sometimes program might get unresponsive and just block. Although I
suspect requests or twitter, I have no idea why. This feature is here to
remedy this.

According to my numbers, it takes between 80 and 160 seconds for the
rate counter to hit. We give it 600 seconds, a much bigger value to be
sure that we are actually resetting the program when it's freezed.